+++ libs/sqlite/src/com/biffweb/sqlite/impl/system.clj
@@ -4,6 +4,7 @@
             [com.biffweb.sqlite.impl.kv :as kv]
             [com.biffweb.sqlite.impl.litestream :as litestream]
             [com.biffweb.sqlite.impl.pool :as pool]
+            [com.biffweb.sqlite.impl.resolver :as resolver]
             [com.biffweb.sqlite.impl.sqldef :as impl.sqldef]))
 
 (defn- start
@@ -50,3 +51,13 @@
      {:biff.sqlite/columns (into {}
                                  (mapcat :biff.sqlite/columns)
                                  @modules-var)})})
+
+(defn schema-module
+  [{:biff.sqlite/keys [extra-init-sql authorize columns]}]
+  {:biff.core/init
+   (fn [_modules-var]
+     {:biff.sqlite/extra-init-sql extra-init-sql
+      :biff.sqlite/authorize      authorize})
+
+   :biff.sqlite/columns  columns
+   :biff.graph/resolvers (resolver/make-resolvers columns)})